+++ storybook/cypress/e2e/document-component.cy.ts
@@ -19,6 +19,11 @@ function expectAlert(text: string | undefined = undefined) {
     .should(text ? "have.text" : "not.exist", text);
 }
 
+const marker: DocumentComponentMarkFunction = (text, source) =>
+  text
+    .split(new RegExp(`(k)`, "gi"))
+    .map((item, index) => (isOdd(index) ? { text: item, highlight: source === "kop" && index === 1 } : item));
+
 describe("Document Component", () => {
   beforeEach(() => {
     cy.visit("http://localhost:45000/iframe.html?id=core-document-component--default");
@@ -26,11 +31,6 @@ describe("Document Component", () => {
   });
 
   it("should mark and highlight", () => {
-    const marker: DocumentComponentMarkFunction = (text, source) =>
-      text
-        .split(new RegExp(`(k)`, "gi"))
-        .map((item, index) => (isOdd(index) ? { text: item, highlight: source === "kop" && index === 1 } : item));
-
     cy.get("@document-component")
       .then(
         ($documentComponent: JQuery<HTMLDsoDocumentComponentElement>) =>
@@ -232,5 +232,24 @@ describe("Document Component", () => {
     expectAlert("Dit onderdeel is vervallen.");
 
     cy.get("@document-component").matchImageSnapshot(`${Cypress.currentTest.title} -- vervallen`);
+  });
+
+  it("marks and highlights alternativeTitle", () => {
+    cy.visit("http://localhost:45000/iframe.html?id=core-document-component--imro")
+      .get("dso-document-component.hydrated")
+      .then(
+        ($documentComponent: JQuery<HTMLDsoDocumentComponentElement>) =>
+          ($documentComponent[0].mark = (text) =>
+            text
+              .split(new RegExp(`(aa)`, "gi"))
+              .map((item, index) => (isOdd(index) ? { text: item, highlight: index === 1 } : item))),
+      )
+      .get("dso-document-component.hydrated")
+      .shadow()
+      .find("#heading-title")
+      .should(
+        "contain.html",
+        'Adequ<mark class="dso-highlight">aa</mark>t <mark>aa</mark>nbod openb<mark>aa</mark>r vervoer',
+      );
   });
 });